refactor(importers): extract shared UK date/amount helpers
Address Octopus review findings #2, #3, #5, #9 on PR #25:

- Extract parseUkDate / parseFloatSafe to src/services/importers/uk-date.ts
  so the Barclays and HSBC parsers share one source of truth.
- parseUkDate now round-trips the parsed date (rejects e.g. 32/13/2026,
  30/02/2026) instead of silently producing a different valid date.
- parseFloatSafe now throws ValidationError with the row number when a
  non-empty cell fails to parse, instead of silently corrupting to 0.
  Empty cells still yield 0 (banks emit empty debit/credit columns).
- Barclays: when both Memo and Subcategory are empty, use a synthetic
  '(no description)' marker so the dedupe hash key stays unique per row
  rather than collapsing all empty-description rows into one bucket.
