+++ src/services/importers/barclays.ts
@@ -1,6 +1,6 @@
 // Barclays CSV parser.
 //
-// TODO: validate against real export.
+// TODO(#26): validate against real export.
 //
 // Best-effort header (Barclays online banking statement download):
 //   Number,Date,Account,Amount,Subcategory,Memo
@@ -12,6 +12,12 @@
 
 import { ValidationError } from '../../lib/errors';
 import { type ParsedTransaction, parseCsv } from './index';
+import { parseFloatSafe, parseUkDate } from './uk-date';
+
+const PARSER_NAME = 'Barclays';
+
+/** Synthetic placeholder when both memo and subcategory are empty. */
+export const EMPTY_DESCRIPTION = '(no description)';
 
 export function parseBarclays(raw: string): ParsedTransaction[] {
   const rows = parseCsv(raw);
@@ -30,7 +36,7 @@ export function parseBarclays(raw: string): ParsedTransaction[] {
     }
   }
   if (headerIdx === -1) {
-    throw new ValidationError('Barclays parser: header row not found');
+    throw new ValidationError(`${PARSER_NAME} parser: header row not found`);
   }
 
   const idx = {
@@ -51,31 +57,16 @@ export function parseBarclays(raw: string): ParsedTransaction[] {
     const sub = idx.subcategory !== -1 ? (row[idx.subcategory] ?? '').trim() : '';
     if (!dateStr) continue;
 
-    const description = memo || sub;
+    // Synthetic marker keeps dedupe keys distinct for rows where both memo and
+    // subcategory are empty. Without this, all such rows on the same date for
+    // the same amount would collapse into one hash.
+    const description = memo || sub || EMPTY_DESCRIPTION;
     out.push({
-      date: parseUkDate(dateStr),
-      amount: parseFloatSafe(amountStr),
+      date: parseUkDate(dateStr, PARSER_NAME),
+      amount: parseFloatSafe(amountStr, PARSER_NAME, i + 1),
       description,
       currency: 'GBP',
     });
   }
   return out;
-}
-
-function parseUkDate(s: string): Date {
-  const m = /^(\d{1,2})\/(\d{1,2})\/(\d{2,4})$/.exec(s);
-  if (!m) throw new ValidationError(`Barclays parser: invalid date '${s}'`);
-  const day = Number.parseInt(m[1] as string, 10);
-  const month = Number.parseInt(m[2] as string, 10);
-  let year = Number.parseInt(m[3] as string, 10);
-  if (year < 100) year += 2000;
-  return new Date(Date.UTC(year, month - 1, day));
-}
-
-function parseFloatSafe(s: string): number {
-  if (!s) return 0;
-  const cleaned = s.replace(/[£,]/g, '').trim();
-  if (!cleaned) return 0;
-  const n = Number.parseFloat(cleaned);
-  return Number.isFinite(n) ? n : 0;
 }
